+++ src/includes/classes/SCore/Base/Wp.php
@@ -67,6 +67,7 @@ class Wp // Stand-alone class.
         } elseif (!($this->template_directory_url_parts = parse_url($this->template_directory_url))) {
             throw new Exception('Failed to parse template directory URL parts.');
         }
-        $this->is_wc_active = defined('WC_VERSION'); // WooCommerce.
+        $this->is_woocommerce_active = defined('WC_VERSION');
+        $this->is_jetpack_active     = defined('JETPACK__VERSION');
     }
 }