+++ src/includes/classes/SCore/Utils/Jetpack.php
@@ -1,0 +1,65 @@
+<?php
+declare (strict_types = 1);
+namespace WebSharks\WpSharks\Core\Classes\SCore\Utils;
+
+use WebSharks\WpSharks\Core\Classes;
+use WebSharks\WpSharks\Core\Interfaces;
+use WebSharks\WpSharks\Core\Traits;
+#
+use WebSharks\Core\WpSharksCore\Classes as CoreClasses;
+use WebSharks\Core\WpSharksCore\Classes\Core\Base\Exception;
+use WebSharks\Core\WpSharksCore\Interfaces as CoreInterfaces;
+use WebSharks\Core\WpSharksCore\Traits as CoreTraits;
+#
+use function assert as debug;
+use function get_defined_vars as vars;
+
+/**
+ * Jetpack utils.
+ *
+ * @since 160720 Jetpack utils.
+ */
+class Jetpack extends Classes\SCore\Base\Core
+{
+    /**
+     * Jetpack markdown class.
+     *
+     * @since 160720 Jetpack utils.
+     *
+     * @param \WPCom_Markdown|null|bool
+     */
+    protected $WPCom_Markdown;
+
+    /**
+     * Markdown via Jetpack.
+     *
+     * @since 160720 Jetpack utils.
+     *
+     * @param string $markdown Markdown.
+     *
+     * @return string HTML from markdown.
+     */
+    public function markdown(string $markdown): string
+    {
+        if (!$this->canMarkdown() || !$this->WPCom_Markdown) {
+            throw $this->c::issue('Jetpack markdown unavailable.');
+        }
+        return (string) $this->WPCom_Markdown->transform($markdown, ['unslash' => false]);
+    }
+
+    /**
+     * Can markdown via Jetpack?
+     *
+     * @since 160720 Jetpack utils.
+     *
+     * @return bool Can markdown?
+     */
+    public function canMarkdown(): bool
+    {
+        if (!isset($this->WPCom_Markdown)) {
+            $this->WPCom_Markdown = $this->Wp->is_jetpack_active && class_exists('WPCom_Markdown')
+                ? \WPCom_Markdown::get_instance() : false;
+        }
+        return (bool) $this->WPCom_Markdown;
+    }
+}